feat: Altered netex file naming strategy
feat: Include check for BODS services after operator search
feat: Created Multioperator tab in My Fares
feat: added operator group in operator settings

fix: Stop users deleting time restrictions used in products
fix: Stop users deleting purchase methods used in products
fix: Stop users deleting passenger types used in products
fix: Adding passenger type card IDs to UI
fix: Changed version of aws amazon ecs deploy task
fix: Deleting products now deletes in S3 products bucket

test: Disabling scheme UI tests and enabling multioperator ones

chore: Added linting rules to types repo

+++ repos/fdbt-site/src/pages/api/deletePassenger.ts
@@ -1,7 +1,15 @@
 import { NextApiResponse } from 'next';
-import { deletePassengerTypeByNocCodeAndId, getGroupPassengerTypeDbsFromGlobalSettings } from '../../data/auroradb';
+import {
+    deletePassengerTypeByNocCodeAndId,
+    getAllProductsByNoc,
+    getGroupPassengerTypeDbsFromGlobalSettings,
+    getPassengerTypeById,
+} from '../../data/auroradb';
 import { redirectToError, redirectTo, getAndValidateNoc } from '../../utils/apiUtils/index';
-import { NextApiRequestWithSession } from '../../interfaces';
+import { ErrorInfo, NextApiRequestWithSession, SinglePassengerType } from '../../interfaces';
+import { getProductsMatchingJson } from '../../data/s3';
+import { updateSessionAttribute } from '../../utils/sessions';
+import { VIEW_PASSENGER_TYPE } from '../../constants/attributes';
 
 export default async (req: NextApiRequestWithSession, res: NextApiResponse): Promise<void> => {
     try {
@@ -25,12 +33,31 @@ export default async (req: NextApiRequestWithSession, res: NextApiResponse): Pro
             throw new Error('This individual cannot be deleted because it is in use in a group.');
         }
 
+        const products = await getAllProductsByNoc(nationalOperatorCode);
+        const matchingJsonLinks = products.map((product) => product.matchingJsonLink);
+        const tickets = await Promise.all(
+            matchingJsonLinks.map(async (link) => {
+                return await getProductsMatchingJson(link);
+            }),
+        );
+
+        const productsUsingPassengerType = tickets.filter((t) => t.passengerType.id === id);
+
+        if (productsUsingPassengerType.length > 0) {
+            const passengerDetails = (await getPassengerTypeById(id, nationalOperatorCode)) as SinglePassengerType;
+            const { name } = passengerDetails;
+            const errorMessage = `You cannot delete ${name} because it is being used in ${productsUsingPassengerType.length} product(s).`;
+            const errors: ErrorInfo[] = [{ id: 'passenger-card-0', errorMessage }];
+            updateSessionAttribute(req, VIEW_PASSENGER_TYPE, errors);
+            redirectTo(res, `/viewPassengerTypes?cannotDelete=${name}`);
+            return;
+        }
+
         await deletePassengerTypeByNocCodeAndId(id, nationalOperatorCode, isGroup === 'true');
 
         redirectTo(res, '/viewPassengerTypes');
     } catch (error) {
         const message = 'There was a problem deleting the selected passenger or group';
-
         redirectToError(res, message, 'api.deletePassenger', error);
     }
 };
